Rename validator to match new naming convention
[#150411707]
