+++ app/javascript/investigations/contacts/ContactFormValidator.js
